Move parameter encoding and validation to `ActionDispatch::Journey::Router`
Parameter encoding and checks to verify whether all encodings are valid
are currently done in the ActionDispatch::Request object. However,
middleware may instantiate a request and access it's parameters. This causes
encoding validation to be performed prior to the controller / action being
known, resulting in an "invalid parameters" error if any of the parameters
are non UTF-8, even if the controller will specify custom encoding or
skip parameter encoding altogether.

This PR moves parameter encoding and validation from the request object
to the router. Once we're in the router, we have the full context on
the controller and action and can properly encode and check the encodings.